requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
flask-smorest==0.44.0
marshmallow==3.21.2
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.3.2
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
alembic==1.13.1
Flask-Cors==4.0.1
pytest-cov==5.0.0
boto3==1.28.39
requests==2.32.3
pdfplumber==0.11.4
pdf2image==1.17.0
pillow==10.4.0
easyocr==1.7.1
torch==2.2.2
numpy==1.26.4
